requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.4
django-allauth==0.61.1
django-cache-url==3.4.5
django-crispy-forms==2.1
environs==11.0.0
gunicorn==21.2.0
idna==3.6
marshmallow==3.21.1
oauthlib==3.2.2
packaging==24.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
